Move exception_records related methods to errors.py
